Created inventory system
Added base functionality for an inventory system. Allows the user to create objects and add item data to those objects. The item data is stored in a data table where the user can modify the items as they wish.
The player character can trace for objects on a specific layer and interact with them. This adds the item to the players inventory and destroys it from the world. Not server replicated.